get_languages/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore >= 2.3.4
flask_cors >= 3.0.10